+++ AirCasting/FeatureFlags/FeatureFlagProvider.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-enum FeatureFlag: Equatable, CaseIterable {
+enum FeatureFlag: String, Equatable, CaseIterable {
     case sdCardSync
     case standaloneMode
 }
